requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy[asyncio]==2.0.41
asyncpg==0.30.0
openai== 1.97.0
celery[redis]==5.5.3
python-dotenv==1.1.1
pydantic-settings==2.10.1
pydantic==2.11.7
tenacity==9.1.2
pydantic[email]
redis==5.2.1
python-jose==3.5.0